backend/ai-strategy/requirements.txt
# AI策略生成服务依赖包

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据处理
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# HTTP客户端
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 配置管理
PyYAML==6.0.1
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 异步支持
aiofiles==23.2.1

# 代码分析和安全检查
ast-decompiler==0.7.0
bandit==1.7.5

# 优化算法（可选，用于高级优化功能）
scipy==1.11.4
# scikit-optimize==0.9.0  # 贝叶斯优化（可选）
# deap==1.4.1  # 遗传算法（可选）

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # 用于测试API

# 开发工具
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# 性能监控（可选）
# prometheus-client==0.19.0
# psutil==5.9.6

# 缓存（可选）
# redis==5.0.1
# aioredis==2.0.1

# 数据库（可选，如果需要持久化）
# sqlalchemy==2.0.23
# alembic==1.13.1
# asyncpg==0.29.0  # PostgreSQL
# aiomysql==0.2.0  # MySQL

# 加密和安全
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# JSON处理
orjson==3.9.10

# 环境变量
python-multipart==0.0.6

# 类型检查
typing-extensions==4.8.0

-r ../../shared/requirements/base.txt
-r ../../shared/requirements/ai.txt